rename structure to networks
